+++ mathlib/Matrix.cpp
@@ -11,7 +11,7 @@ Matrix::Matrix(int rows, int cols) :
 {
 }
 
-Matrix::Matrix(int rows, int cols, std::vector<double> vec) :
+Matrix::Matrix(int rows, int cols, std::vector<float> vec) :
     rows(rows), cols(cols), mat(std::move(vec))
 {
     if (mat.size() != rows * cols)
@@ -30,7 +30,7 @@ const int Matrix::getCols() const noexcept
     return cols;
 }
 
-const std::vector<double>& Matrix::getVector() const noexcept
+const std::vector<float>& Matrix::getVector() const noexcept
 {
     return mat;
 }
@@ -41,7 +41,7 @@ void Matrix::setDimensions(int rows, int cols)
     this->cols = cols;
 
     mat.resize(rows * cols);
-    std::fill(mat.begin(), mat.end(), 0);
+    std::fill(mat.begin(), mat.end(), 0.f);
 }
 
 void Matrix::print() const
@@ -49,7 +49,7 @@ void Matrix::print() const
     std::cout << *this;
 }
 
-Matrix Matrix::applyFunc(std::function<double(double)> func) const
+Matrix Matrix::applyFunc(std::function<float(float)> func) const
 {
     const Matrix& m = *this;
     Matrix mo(m.rows, m.cols);
@@ -79,37 +79,37 @@ Matrix Matrix::T() const
     return tm;
 }
 
-double Matrix::sum() const
+float Matrix::sum() const
 {
     return std::accumulate(mat.begin(), mat.end(), 0.0);
 }
 
-std::vector<double>::iterator Matrix::begin()
+std::vector<float>::iterator Matrix::begin()
 {
     return mat.begin();
 }
 
-std::vector<double>::iterator Matrix::end()
+std::vector<float>::iterator Matrix::end()
 {
     return mat.end();
 }
 
-double& Matrix::operator()(int row, int col)
+float& Matrix::operator()(int row, int col)
 {
     return mat.at(row * cols + col); // change to [] in future
 }
 
-double Matrix::operator()(int row, int col) const
+float Matrix::operator()(int row, int col) const
 {
     return mat.at(row * cols + col); // change to [] in future
 }
 
-double& Matrix::operator[](int i)
+float& Matrix::operator[](int i)
 {
     return mat.at(i); // change to [] in future
 }
 
-double Matrix::operator[](int i) const
+float Matrix::operator[](int i) const
 {
     return mat.at(i); // change to [] in future
 }
